backend/requirements.txt
langchain
fastapi
pydantic
supabase
python-dotenv
uvicorn
langchain_community
langchain_google_genai
requests
